Refactor project structure and update dependencies
- Moved main entry point to './src/main/index.js' and updated related paths in 'forge.config.js'.
- Replaced './src/index.html' and './src/renderer.js' with new paths in 'forge.config.js'.
- Removed unused files: 'src/index.css', 'src/index.html', 'src/main.js', 'src/preload.js', and 'src/renderer.js'.
- Added 'sqlite3' as a new dependency in 'package.json' and updated 'package-lock.json' accordingly.
