+++ forge.config.js
@@ -37,11 +37,11 @@ module.exports = {
           config: './webpack.renderer.config.js',
           entryPoints: [
             {
-              html: './src/index.html',
-              js: './src/renderer.js',
+              html: './src/renderer/index.html',
+              js: './src/renderer/renderer.js',
               name: 'main_window',
               preload: {
-                js: './src/preload.js',
+                js: './src/main/preload.js',
               },
             },
           ],
