+++ package.json
@@ -35,6 +35,7 @@
     "style-loader": "^3.3.4"
   },
   "dependencies": {
-    "electron-squirrel-startup": "^1.0.1"
+    "electron-squirrel-startup": "^1.0.1",
+    "sqlite3": "^5.1.7"
   }
 }
