+++ src/main/services/crypto-engine.js
@@ -1,0 +1,4 @@
+// Placeholder for crypto engine implementation
+module.exports = {};
+
+
