+++ src/main/services/database-manager.js
@@ -1,0 +1,4 @@
+// Placeholder for database manager implementation
+module.exports = {};
+
+
