+++ src/renderer/index.html
@@ -9,4 +9,6 @@
     <h1>💖 Hello World!</h1>
     <p>Welcome to your Electron application.</p>
   </body>
-</html>
+  </html>
+
+
