+++ src/renderer/renderer.js
@@ -1,0 +1,5 @@
+import './styles/main.css';
+
+console.log('👋 This message is being logged by "renderer.js", included via webpack');
+
+
