+++ src/renderer/styles/main.css
@@ -5,3 +5,5 @@ body {
   max-width: 38rem;
   padding: 2rem;
 }
+
+
