+++ src/shared/constants.js
@@ -1,0 +1,3 @@
+// Shared constants between main and renderer processes can live here.
+
+
